requirements.txt
glfw
PyOpenGL
numpy
pyglm
Pillow
